Use Auth with policy

+++ app/Http/Controllers/QuestionsController.php
@@ -8,6 +8,11 @@ use App\Http\Requests\AskQuestionRequest;
 
 class QuestionsController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -62,9 +67,12 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        /*
         if (\Gate::denies('update-question', $question)) {
             abort(403, "Access denied");
         }
+        */
+        $this->authorize('update', $question);
         return view('questions.edit', compact('question'));
     }
 
@@ -77,9 +85,12 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        /*
         if (\Gate::denies('update-question', $question)) {
             abort(403, "Access denied");
         }
+        */
+        $this->authorize('update', $question);
         $question->update($request->only('title', 'body'));
         return redirect()->route('questions.index')->with('success', 'Your question has been updated');
     }
@@ -92,9 +103,12 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        /*
         if (\Gate::denies('delete-question', $question)) {
             abort(403, "Access denied");
         }
+        */
+        $this->authorize('delete', $question);
         $question->delete();
         return redirect()->route('questions.index')->with('success', 'Your question has been deleted');
     }
